Time: 94 ms (23.31%), Space: 33.1 MB (10.18%) - LeetHub

+++ 2244-number-of-laser-beams-in-a-bank/2244-number-of-laser-beams-in-a-bank.cpp
@@ -9,9 +9,9 @@ public:
             }
         }
 
-        long long int ans = 0;
+        int ans = 0;
         for(int i=1;i<v.size();i++){
-            ans += v[i] * v[i-1];
+            ans+=v[i]*v[i-1];
         }
 
         return ans;
